fixes for ramsons okapi spitfire

+++ cypress/e2e/marc/marc-bibliographic/create-new-marc-bib/manual-linking/linking-repeatable-fields-when-creating-marcBib.cy.js
@@ -54,7 +54,7 @@ describe('MARC', () => {
             rowIndex: 5,
             value: 'C380738 Good and evil History',
             tag: 650,
-            boxFourth: '$a C380738 Good and evil $x History',
+            boxFourth: '$a C380738 Good and evil',
             boxFifth: '',
             boxSixth: '$0 http://id.loc.gov/authorities/subjects/sh2009125989',
             boxSeventh: '',
@@ -119,7 +119,9 @@ describe('MARC', () => {
 
         after('Delete created user and data', () => {
           cy.getAdminToken();
-          Users.deleteViaApi(userData.userId);
+          if (userData?.userId) {
+            Users.deleteViaApi(userData.userId);
+          }
           for (let i = 0; i < 2; i++) {
             MarcAuthority.deleteViaAPI(createdAuthorityIDs[i], true);
           }
@@ -188,7 +190,7 @@ describe('MARC', () => {
             InventoryInstance.viewSource();
             InventoryViewSource.verifyAbsenceOfValue(linkingTagAndValues[1].value);
             InventoryViewSource.contains(
-              'Linked to MARC authority\n\t650\t   \t$a C380738 Good and evil $x History $0 http://id.loc.gov/authorities/subjects/sh2009125989 $9',
+              'Linked to MARC authority\n\t650\t   \t$a C380738 Good and evil $0 http://id.loc.gov/authorities/subjects/sh2009125989 $9',
             );
           },
         );
